Display active filters regardless of the filter panel
